Move Liftable to crate::abstract module
The `Liftable` trait is its own thing, it defines a trait to lift
miniscript, descriptors, and concrete policies to an abstract policy.

Move the `Liftable` and associated error type and impls to a new `lift`
module that lives under the recently created `crate::r#abstract` module.

Note we alse remove the `Semantic` and `Concrete` public
re-names/re-exports from `crate::policy` because we are transitioning
away from using the term "semantic", however to minimise the changes in
this patch we add a bunch of

  `use crate::policy::concrete::Policy as Concrete`

statements.
