+++ fuzz/fuzz_targets/compile_descriptor.rs
@@ -1,11 +1,11 @@
 use std::str::FromStr;
 
 use honggfuzz::fuzz;
+use miniscript::r#abstract::Liftable;
 use miniscript::{policy, Miniscript, Segwitv0};
-use policy::Liftable;
 
 type Script = Miniscript<String, Segwitv0>;
-type Policy = policy::Concrete<String>;
+type Policy = policy::concrete::Policy<String>;
 
 fn do_test(data: &[u8]) {
     let data_str = String::from_utf8_lossy(data);
