+++ fuzz/fuzz_targets/roundtrip_concrete.rs
@@ -4,7 +4,7 @@ use honggfuzz::fuzz;
 use miniscript::policy;
 use regex::Regex;
 
-type Policy = policy::Concrete<String>;
+type Policy = policy::concrete::Policy<String>;
 
 fn do_test(data: &[u8]) {
     let data_str = String::from_utf8_lossy(data);
