+++ fuzz/fuzz_targets/roundtrip_semantic.rs
@@ -3,7 +3,7 @@ use std::str::FromStr;
 use honggfuzz::fuzz;
 use miniscript::policy;
 
-type Policy = policy::Semantic<String>;
+type Policy = policy::r#abstract::Policy<String>;
 
 fn do_test(data: &[u8]) {
     let data_str = String::from_utf8_lossy(data);
