+++ src/abstract/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: CC0-1.0
+
+//! WIP - TODO: Describe this module.
+
+pub use self::lift::{LiftError, Liftable};
+
+mod lift;
